refactor: work with databinding and model WIP
