+++ src/const.js
@@ -1,0 +1,26 @@
+const Mode = {
+  DEFAULT: 'DEFAULT',
+  EDITING: 'EDITING'
+};
+
+const SortType = {
+  DAY: 'day',
+  EVENT: 'eventSort',
+  TIME: 'time',
+  PRICE: 'price',
+  OFFERS: 'offers'
+};
+
+const UserAction = {
+  UPDATE_POINT: 'UPDATE_TASK',
+  ADD_POINT: 'ADD_POINT',
+  DELETE_POINT: 'DELETE_POINT'
+};
+
+const UpdateType = {
+  PATCH: 'PATCH',
+  MINOR: 'MINOR',
+  MAJOR: 'MAJOR'
+};
+
+export {Mode, SortType, UserAction, UpdateType };
